+++ src/main/kotlin/api/vkPlatform.kt
@@ -2,7 +2,6 @@ package api
 
 import api.objects.VkUser
 import com.google.gson.Gson
-import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import log
@@ -37,17 +36,17 @@ class VkPlatform : PlatformApiInterface {
         return profiles.map { gson.fromJson(it, VkUser::class.java) }
     }
 
-    override fun getUserIdByName(showingName: String): Int? {
+    override fun getUserIdByName(showingName: String): Long? {
         val resp = post(
                 "users.get", mutableMapOf(
                 "user_ids" to showingName
             )
         )
         val json = resp?.asJsonObject
-        return json?.get("response")?.asJsonArray?.get(0)?.asJsonObject?.get("id")?.asInt
+        return json?.get("response")?.asJsonArray?.get(0)?.asJsonObject?.get("id")?.asLong
     }
 
-    override fun getUserNameById(id: Int): String? {
+    override fun getUserNameById(id: Long): String? {
         val resp = post("users.get", mutableMapOf(
                 "user_ids" to id,
                 "fields" to "screen_name"
@@ -56,14 +55,14 @@ class VkPlatform : PlatformApiInterface {
         return json?.get("response")?.asJsonArray?.get(0)?.asJsonObject?.get("screen_name")?.asString
     }
 
-    override fun kickUserFromChat(chatId: Int, userId: Int) {
+    override fun kickUserFromChat(chatId: Long, userId: Long) {
         post("messages.removeChatUser", mutableMapOf(
                 "chat_id" to chatId - 2000000000,
                 "user_id" to userId
         ))
     }
 
-    override fun send(text: String, chatId: Int, urls: List<String>) {
+    override fun send(text: String, chatId: Long, urls: List<String>) {
         val attachments = mutableListOf<String>()
         for (url in urls) {
             val imageConnection = URL(url).openConnection()
@@ -76,7 +75,7 @@ class VkPlatform : PlatformApiInterface {
     }
 
     @OptIn(ExperimentalStdlibApi::class)
-    fun uploadPhoto(peer_id: Int, data: ByteArray): String? {
+    fun uploadPhoto(peer_id: Long, data: ByteArray): String? {
         val serverData = post(
             "photos.getMessagesUploadServer",
             mutableMapOf(
